ml-pipeline/pyproject.toml
[project]
name = "ml-pipeline"
version = "0.1.0"
description = "MLOps Pipeline for Machine Learning Model Training and Evaluation"
authors = [
    {name = "MLOps Team"}
]
readme = "README.md"
packages = [
    {include = "ml_pipeline", from = "src"}
]

requires-python = ">=3.9,<3.12"
dependencies = [
    "torch (==2.1.2)",
    "lightning (==2.5.5)",
    "numpy (==1.24.3)",
    "pandas (==2.1.4)",
    "scikit-learn (==1.3.2)",
    "psycopg2-binary (==2.9.9)",
    "sqlalchemy (==2.0.23)",
    "mlflow (==2.8.1)",
    "pyyaml (==6.0.1)",
    "kfp (==2.4.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-cov = "^7.0.0"
black = "^25.9.0"
flake8 = "^7.3.0"
mypy = "^1.18.2"

